requirements.txt
jupyterlab>3.0
transformers>4.0
torch>2.4
python-dotenv>0.15
accelerate>=0.26.0
Flask
pandas>1.0
Flask_Cors>=5.0
PyPDF2
requests
pandas
pydub 
vosk 
argparse
presidio-image-redactor>=2.2.33
Pillow>=10.0.0
pytesseract>=0.3.10
spacy>=3.0.0 
easyocr
